refactor(feature/#37): CustomChart 컴포넌트 만들어서 분리
